[refactor] use widget
